--- pkg/apis/query/v0alpha1/response.go
@@ -1,15 +1,0 @@
-package v0alpha1
-
-import (
-	"github.com/grafana/grafana-plugin-sdk-go/backend"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-// Wraps backend.QueryDataResponse, however it includes TypeMeta and implements runtime.Object
-// +k8s:deepcopy-gen:interfaces=k8s.io/apimachinery/pkg/runtime.Object
-type QueryDataResponse struct {
-	metav1.TypeMeta `json:",inline"`
-
-	// Backend wrapper (external dependency)
-	backend.QueryDataResponse `json:",inline"`
-}
